Temporarily disabling orphaned events deletion
Currently we check for ALL events in both main and archived orgmode
file. This is good, as we can update events that ended up in archive
before they were updated in the calendar, however this will take more
and more time to sync as far as we add more and more tasks in orgmode.
To avoid that, I could
1. make deletion optional, or guarded with manual confirm (which is kind
   of boring)
2. remove archive.org file from the watch list

while (2) is much better, the current implementation would delete a
bunch of events that exists in the archive.md.
So, for now, I just skip the deletion part